refactor: reorganize prompt utils into separate modules
- Split loadYaml and mergeDicts into new agentLoad module
- Move XML-related functions from promptUtils to xmlUtils
- Add new XML utility functions for CDATA operations
