+++ coauthor-for-vs-code/src/utils/xmlUtils.ts
@@ -1,3 +1,94 @@
+import { debug, error, initializeLogging } from '../logger/logUtils';
+import { readFile } from './fileUtils';
+
+const CHANNEL = 'XMLUtils';
+initializeLogging(CHANNEL);
+
+/**
+ * Get XML formatted string from a single file
+ * @param file Path to the file
+ * @returns XML formatted string containing file content
+ */
+export async function getXmlFormatFromFile(file: string): Promise<string> {
+  try {
+    const content = await readFile(file);
+    return `<document name="${file}">\n${content}\n</document>`;
+  } catch (err) {
+    error(
+      CHANNEL,
+      `Error formatting file as XML: ${err instanceof Error ? err.message : String(err)}`,
+    );
+    throw err;
+  }
+}
+
+/**
+ * Get XML formatted string from multiple files
+ * @param files List of file paths
+ * @returns XML formatted string containing all file contents, or null if no files
+ */
+export async function getXmlFormatFromFiles(
+  files: string[],
+): Promise<string | null> {
+  try {
+    if (!files || files.length === 0) {
+      return null;
+    }
+
+    const xmlPromises = files.map((file) => getXmlFormatFromFile(file));
+    const xmlContents = await Promise.all(xmlPromises);
+    return xmlContents.join('\n');
+  } catch (err) {
+    error(
+      CHANNEL,
+      `Error formatting files as XML: ${err instanceof Error ? err.message : String(err)}`,
+    );
+    throw err;
+  }
+}
+
+/**
+ * Wrap content of specified tags with CDATA sections
+ */
+export function addCdataToTags(xmlData: string, tags: string[]): string {
+  try {
+    return tags.reduce((result, tag) => {
+      const pattern = new RegExp(`(<${tag}>)(.*?)(</${tag}>)`, 'gs');
+      return result.replace(pattern, '$1<![CDATA[$2]]>$3');
+    }, xmlData);
+  } catch (err) {
+    error(
+      CHANNEL,
+      `Error adding CDATA to tags: ${err instanceof Error ? err.message : String(err)}`,
+    );
+    throw err;
+  }
+}
+
+/**
+ * Wrap content of specified tags with CDATA sections, handling attributes
+ */
+export function addCdataToTagsMultiple(
+  xmlData: string,
+  tags: string[],
+): string {
+  try {
+    return tags.reduce((result, tag) => {
+      const pattern = new RegExp(
+        `(<${tag}(?:\\s+[^>]*)?>)(.*?)(</${tag}>)`,
+        'gs',
+      );
+      return result.replace(pattern, '$1<![CDATA[$2]]>$3');
+    }, xmlData);
+  } catch (err) {
+    error(
+      CHANNEL,
+      `Error adding CDATA to tags with attributes: ${err instanceof Error ? err.message : String(err)}`,
+    );
+    throw err;
+  }
+}
+
 export function extractTextFromTags(
   inputContent: string,
   documentTag: string,
@@ -5,4 +96,26 @@ export function extractTextFromTags(
   const regex = new RegExp(`<${documentTag}>(.*?)<\/${documentTag}>`, 's');
   const match = inputContent.match(regex);
   return match ? match[1] : inputContent;
+}
+
+/**
+ * Remove specified XML tags and their content from input string
+ */
+export function filterTagsFromText(
+  content: string,
+  tags: string | string[],
+): string {
+  try {
+    const tagArray = typeof tags === 'string' ? [tags] : tags;
+    return tagArray.reduce((result, tag) => {
+      const pattern = new RegExp(`<${tag}>.*?</${tag}>\\s*`, 'gs');
+      return result.replace(pattern, '');
+    }, content);
+  } catch (err) {
+    error(
+      CHANNEL,
+      `Error filtering tags from text: ${err instanceof Error ? err.message : String(err)}`,
+    );
+    throw err;
+  }
 }
